refactor(backend): organize bn254 crate into scheme/frontend/solver modules
The flat module layout collided the prove/verify orchestration traits with
the prover/verifier scheme structs (prove.rs vs prover.rs, verify.rs vs
verifier.rs) and mixed four distinct concerns in one namespace.

Group them: scheme/ (serialized PKP/PKV/proof types + file-format glue),
frontend/ (Noir/ACIR bridge: witness generation, ABI, input parsing, the
noir_to_native conversion), solver/ (witness solving + the fe_convert
bridges), and prove.rs/verify.rs at the root for orchestration.

Pure reorganization, no logic change. lib.rs re-exports keep the public API
flat, so no consumer is touched and the prove/prover collision is resolved
by module path. 72/72 e2e green.
