+++ provekit/backend/bn254/src/frontend/native.rs
@@ -1,0 +1,12 @@
+use {
+    acir::FieldElement as NoirElement,
+    ark_ff::{BigInt, PrimeField},
+    provekit_common::FieldElement,
+};
+
+/// Convert a Noir field element to a native `FieldElement`.
+#[inline(always)]
+pub fn noir_to_native(n: NoirElement) -> FieldElement {
+    let limbs = n.into_repr().into_bigint().0;
+    FieldElement::from(BigInt(limbs))
+}
