+++ provekit/backend/bn254/src/frontend/witness_generator.rs
