+++ provekit/backend/bn254/src/solver/digits.rs
