--- provekit/backend/bn254/src/witness/mod.rs
@@ -1,4 +1,0 @@
-mod digits;
-mod limb_io;
-mod ram;
-pub(crate) mod witness_builder;
